+++ authlib/admin_oauth/views.py
@@ -3,36 +3,20 @@ import re
 from django.conf import settings
 from django.contrib import auth, messages
 from django.shortcuts import redirect
-from django.utils.http import is_safe_url
 from django.utils.translation import ugettext as _
 from django.views.decorators.cache import never_cache
 
 from authlib.google import GoogleOAuth2Client
+from authlib.views import set_next_cookie, retrieve_next
 
 
-REDIRECT_SESSION_KEY = "admin-oauth-next"
 ADMIN_OAUTH_PATTERNS = settings.ADMIN_OAUTH_PATTERNS
 
 
-def retrieve_next(request):
-    next = request.session.pop(REDIRECT_SESSION_KEY, None)
-    return (
-        next
-        if is_safe_url(
-            url=next,
-            allowed_hosts={request.get_host()},
-            require_https=request.is_secure(),
-        )
-        else None
-    )
-
-
 @never_cache
+@set_next_cookie
 def admin_oauth(request):
     client = GoogleOAuth2Client(request)
-
-    if request.GET.get("next"):
-        request.session[REDIRECT_SESSION_KEY] = request.GET["next"]
 
     if all(key not in request.GET for key in ("code", "oauth_token")):
         return redirect(client.get_authentication_url())